# self-version/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "A simple CRUD API for tasks with Supabase authentication"  },
    "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
    "paths": {
      "/auth/signup": {
      "post": {
        "summary": "Sign up a new user",
        "description": "Registers a new user with email and password using Supabase Auth.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "user@example.com" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created successfully" },
          "400": { "description": "Invalid input or signup error" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Log in a user",
        "description": "Authenticates a user and returns access and refresh tokens.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "user@example.com" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": { "type": "string" },
                    "refresh_token": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Log out a user",
        "description": "Ends the current user session and invalidates the token.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Logout successful, no content returned" },
          "401": { "description": "Invalid or missing token" }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Get authenticated user profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Returns user profile data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "email": { "type": "string" },
                    "created_at": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid or missing token" }
        }
      }
    },
    "/tasks": {
      "get": {
    "summary": "List all tasks",
    "responses": {
      "200": {
        "description": "Successful response",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "title": { "type": "string" },
                  "done": { "type": "boolean" }
                }
              }
            }
          }
        }
      }
    }
  },
  "post": {
    "summary": "Create a new task",
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "title": { "type": "string" }
            },
            "required": ["title"]
          }
        }
      }
    },
    "responses": {
      "201": {
        "description": "Task created successfully",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "id": { "type": "integer" },
                "title": { "type": "string" },
                "done": { "type": "boolean" }
              }
            }
          }
        }
      },
      "400": {
        "description": "Invalid input"
      }
    }
  }
},

    "/tasks/{id}": {
  "parameters": [
    {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
        "type": "integer"
      },
      "description": "ID of the task"
    }
  ],
  "get": {
    "summary": "Get a single task",
    "responses": {
      "200": {
        "description": "Task found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "id": { "type": "integer" },
                "title": { "type": "string" },
                "done": { "type": "boolean" }
              }
            }
          }
        }
      },
      "404": { "description": "Task not found" }
    }
  },
  "put": {
    "summary": "Update a task",
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "title": { "type": "string" },
              "done": { "type": "boolean" }
            }
          }
        }
      }
    },
    "responses": {
      "200": { "description": "Task updated successfully" },
      "400": { "description": "Invalid input" },
      "404": { "description": "Task not found" }
    }
  },
  "delete": {
    "summary": "Delete a task",
    "responses": {
      "204": { "description": "Task deleted successfully" },
      "404": { "description": "Task not found" }
    }
  }
}

  }
}